BLEN-264: Basic integration MaterialX nodes into Blender shader nodetree
