+++ materialx/bl_nodes/output.py
@@ -1,7 +1,8 @@
 # SPDX-License-Identifier: GPL-2.0-or-later
 # Copyright 2022, AMD
+import MaterialX
 
-from .node_parser import NodeParser, Id
+from .node_parser import NodeParser, Id, log
 
 
 class ShaderNodeOutputMaterial(NodeParser):
@@ -15,14 +16,12 @@ class ShaderNodeOutputMaterial(NodeParser):
         if surface is None:
             return None
 
-        if surface.type == 'BSDF':
-            surface = self.create_node('surface', 'surfaceshader', {
-                'bsdf': surface,
-            })
-        elif surface.type == 'EDF':
-            surface = self.create_node('surface', 'surfaceshader', {
-                'edf': surface,
-            })
+        linked_input_type = surface.getType() if isinstance(surface, MaterialX.Node) else surface.type
+
+        if linked_input_type != 'surfaceshader':
+            log.warn("Incorrect node tree to export: output node doesn't have correct input")
+
+            return None
 
         result = self.create_node('surfacematerial', 'material', {
             'surfaceshader': surface,
